backend-ai/requirements.txt
flask
flask-cors
ultralytics
pillow
pytest
